Add dialogs depending on number of telescope items for clarity
This particularly targets the dialog after the bat puzzle and the dialog when returning to Beans' room after visiting the bat (both now split in two).
Warning: This takes care of the common "sequence break" of doing the chest puzzle in Beans' room first. However, confusion is still possible if people then do the bat puzzle before the library puzzle. This situation would be too hard to fix with the current game code.